// generated/src/app/api-models/cieb-web-code-decode.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebWebCodeDecode {
  userDefined6: string;
  userDefined5: string;
  userDefined4: string;
  userDefined3: string;
  updateProcess: string;
  updateUser: string;
  updateDatetime: Date;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  userDefined2: string;
  userDefined1: string;
  decode4: string;
  decode3: string;
  decode2: string;
  decode1: string;
  code: string;
  codeType: string;
  seqCodeId: number;
}